Move StateGen to own file, and ran make format.
